+++ src/main/java/com/emazon/api/stock/domain/exception/BrandNullPointerException.java
@@ -1,0 +1,8 @@
+package com.emazon.api.stock.domain.exception;
+
+public class BrandNullPointerException extends  RuntimeException{
+
+    public BrandNullPointerException(String message) {
+        super(message);
+    }
+}
